% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_variables.R
\name{assign_variables}
\alias{assign_variables}
\title{Assign all columns in a 1-row data.frame to a global variable for use in simulations}
\usage{
assign_variables(df)
}
\arguments{
\item{df}{A one-row data frame object}
}
\value{
Nothing
}
\description{
Assign all columns in a 1-row data.frame to a global variable for use in simulations
}
