% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ROC-plot.r
\name{AUC}
\alias{AUC}
\title{Calculate the Area under an ROC curve}
\usage{
AUC(x, case)
}
\arguments{
\item{x}{A vector of marker values}

\item{case}{A logical vector where TRUE indicates that the corresonding marker value in X}
}
\description{
Calculate the Area under an ROC curve.
}
\examples{
AUC( x = rnorm(100) + rep(c(1,0),50), case = rep(c(TRUE,FALSE),50))
}

