% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-misc.r
\name{axis}
\alias{axis}
\title{Add an Axis to a Plot.}
\usage{
axis(..., log = FALSE)
}
\arguments{
\item{...}{other \link{graphical parameters} may also be passed as
    arguments to this function, particularly, \code{cex.axis},
    \code{col.axis} and \code{font.axis} for axis annotation, \code{mgp}
    and \code{xaxp} or \code{yaxp} for positioning, \code{tck} or
    \code{tcl} for tick mark length and direction, \code{las} for
    vertical/horizontal label orientation, or \code{fg} instead of
    \code{col}, and \code{xpd} for clipping.  See \code{\link{par}} on these.

    Parameters \code{xaxt} (sides 1 and 3) and \code{yaxt} (sides 2 and
    4) control if the axis is plotted at all.

    Note that \code{lab} will partial match to argument
    \code{labels} unless the latter is also supplied.  (Since the
    default axes have already been set up by \code{\link{plot.window}},
    \code{lab} will not be acted on by \code{axis}.)}

\item{log}{If set to \code{TRUE}, displays the tic marks on the log scale (default=FALSE)}
}
\description{
Add an Axis to a Plot. A wrapper for \code{\link[graphics]{axis}}
which supports the ability to log transforming the default axis labels
which comes in handy when calling plot(log(x),log(y),...) is easier than
calling  plot(x,y,log='xy',...)
}
\seealso{
the original \code{\link[graphics]{axis}} function
}

