% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ROC-plot.r
\name{bev}
\alias{bev}
\title{Plot a Beysian Expected Value curve}
\usage{
bev(x, case, conf.level = 0.95, band = TRUE, ci = TRUE, col = "black",
  col.band = "colorize", add = FALSE, lwd = 3, xlim = 0:1, ylim = 0:1,
  xlab = "Estimated False Positve Rate",
  ylab = "Estimated True Positive Rate", ...)
}
\arguments{
\item{x}{A vector of marker values}

\item{case}{A logical vector where TRUE indicates that the corresonding marker value in X
is taken from a 'Case', and FALSE otherwise.}

\item{conf.level}{confidence level of the interval. (default = 0.95)}

\item{band}{logical. Should a confidence band be plotted?}

\item{ci}{If TRUE, confidence bands are plotted. (default = TRUE)}

\item{col}{The color of the BEV curve. (default = 'black')}

\item{col.band}{The color of the confidenc band, or 'colorize' if a color scale is used. (default = 'colorize')}

\item{add}{If TRUE, the ROC curve is added to an existing plot. (default = FALSE)}

\item{lwd}{Line width of the ROC curve. (default = 3)}

\item{xlim,ylim}{bounds for the x and y axes}

\item{xlab}{x-axis label (default = 'Estimated False Positive Rate')}

\item{ylab}{y-axis label (default = 'Estimated True Positive Rate')}

\item{...}{additional arguments plot()}
}
\description{
Plot a Beysian Expected Value curve, which is similar to an ROC curve,
though maximum likelihood esimtates of the true and false positive rates
are used instead of the (biased) true and false positivity rates. In addition
confidence bands are included  beta distribution, which is the prior for
the binomial distribution
}

