\name{blanks}
\alias{blanks}
\title{Dreate blank strings with a given number of characters}
\usage{
  blanks(n)
}
\arguments{
  \item{n}{A numeric vector (number of times to duplicate)}
}
\value{
  A character vector
}
\description{
  Dreate blank strings with a given number of characters
}
\seealso{
  Examples for \code{\link{regmatches}}
}

