% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\name{cnum}
\alias{cnum}
\title{Converts numbers into their string ordinals}
\usage{
cnum(x)
}
\arguments{
\item{x}{an integer vector}
}
\description{
Converts numbers into their string ordinals.
}
\examples{
cnum(1:5)
}

