\name{Collection-class}
\alias{Collection-class}
\title{Collection-class}
\usage{
  new_Collection(...)
}
\description{
  A vector of S4 objects with a shared environment to
  contain metadata.
}
\seealso{
  \code{\link{makeCollectionClass}},
  \code{\link{collectionConstructor}},
  \code{\link{collectionValidator}},
  \code{\link{collectionShower}}.
}

