\name{collectionConstructor}
\alias{collectionConstructor}
\title{Create a constructor function for classes extending \code{list}}
\usage{
  collectionConstructor(Class)
}
\arguments{
  \item{Class}{Name of the collection class.}
}
\description{
  Create a constructor function for classes extending
  \code{list}
}

