% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.r
\name{colorchoose}
\alias{colorchoose}
\title{Choose from among the named colors in R}
\usage{
colorchoose(n = 1, alpha, term = FALSE)
}
\arguments{
\item{n}{the number of colors to choose}

\item{alpha}{(optional) an integer between 1 and 255 with the alpha chanel}

\item{term}{logical, should colors be limited to the 26 terminal colors}
}
\description{
Choose from among the named colors in R using a graphical interface
}

