% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Compose}
\alias{Compose}
\alias{\%.\%}
\title{Compose functions}
\usage{
Compose(...)

g \%.\% f
}
\arguments{
  \item{\dots}{The functions to be composed.}

  \item{fn1,fn2}{Two functions to compose (infix notaion)}
}
\description{
Returns a function that applies the last argument to its input, than
the penultimate argument and so on.
}
\examples{
x <- c(1,1,2,2,3,3)
 nunique <- Compose(length, unique)
 nunique(x) == length(unique(x))

 ## usefull in lapply constructs
 sapply(mtcars, length \%.\% unique)
}

