% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dapply.r
\name{dapply}
\alias{dapply}
\title{Apply a Function Over a Ragged Array in a data.frame}
\usage{
dapply(X, INDEX, FUN, ...)
}
\arguments{
\item{X}{an atomic object, typically a vector.}

\item{INDEX}{list of one or more factors, each of same length as
\code{X}.  The elements are coerced to factors by \code{\link{as.factor}}.}

\item{FUN}{the function to be applied, or \code{NULL}.  In the case of
functions like \code{`+`}, \code{\%*\%}, etc., the function name must
be backquoted or quoted.  If \code{FUN} is \code{NULL}, tapply
returns a vector which can be used to subscript the multi-way array
\code{tapply} normally produces.}

\item{...}{optional arguments to \code{FUN}: the Note section.}
}
\description{
Apply a Function Over a Ragged Array in a data.frame (D is for data.frame)
Like \code{\link[base]{sapply}}, tries to coerse the return values into a vector.
The return values of the function FUN inserted into the output vector
in posisions corresponding to the values of INDEX
}
\details{
Note that the \code{FUN} must return values of length 1 or with length equal that
of it's first input.
}

