% Generated by roxygen2 (4.0.0): do not edit by hand
\name{db_count}
\alias{db_count}
\title{Count rows in a db table}
\usage{
db_count(con, tbl)
}
\arguments{
  \item{con}{a connection object.}

  \item{tbl}{name of table in database.}
}
\description{
Count rows in a db table
}
\keyword{internal}

