% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/density-plot.r
\name{density.group}
\alias{density.group}
\title{Plot smooth densities of marker values by group}
\usage{
\method{density}{group}(group, y, col = "skyblue2", border = "dodgerblue4",
  labels = if (inherits(group, "factor")) levels(group) else unique(group),
  xlab = deparse(substitute(group)), ylab = deparse(substitute(y)),
  axes = TRUE, log = "", bg.panel = "white", cex.xlab = 1, type = "p",
  kernel = "c", adjust = 0.5, maxwidth = 0.4, bandwidth,
  med.text = TRUE, xlim, ...)
}
\arguments{
\item{group}{Group membership}

\item{y}{Values to be plotted}

\item{col}{fill color of the density blobs (default = "skyblue2", optionally a vector of colors)}

\item{border}{border color of the density blobs (default = "dodgerblue4", optionally a vector of colors)}

\item{labels}{= if(inherits(group,'factor')) levels(group) else unique(group)}

\item{xlab,ylab}{x and y axis lables.}

\item{axes}{Axes to be plotted ('x','y','xy', or logical; default = 'xy')}

\item{log}{a character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic and "xy" or "yx" if both axes are to be logarithmic.}

\item{bg.panel}{Colors used as backgrounds for the panels}

\item{cex.xlab}{Character expansion factor for xlab (default = 1)}

\item{type}{Type of plot (default = 's'; 's' is the only supported option for now)}

\item{kernel}{kernal type (devault = 'c'; see \code{\link[stats]{density}} for details}

\item{adjust}{adjustment parmeter. See \code{\link[stats]{density}} for details.}

\item{maxwidth}{= 0.4 maximum half-width of the density blob}

\item{bandwidth}{bandwidth smoothing parameter. See \code{\link[stats]{density}} for details.}

\item{med.text}{If TRUE, the word 'median' is included above the meedian line.}

\item{xlim}{limitis of the x axis.  See \code{link[graphics]{plot.default}} for details.}

\item{...}{additional arguments to plot()}
}
\description{
Plot smooth densities of marker values by group.
}
\examples{
density.group(group = rep(c('small','medium','large'),times= 5),
y = rep(1:3,times = 5) + rnorm(15),
col = c('blue','green','red'),
labels = c('large','medium','small'),
xlab = 'group',
ylab = 'y value',
main = 'sample group plot',
log = '',
bg.panel = c('white','gray') )
}

