% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date-functions.r
\name{dtos}
\alias{dtos}
\title{Date to string conversions}
\usage{
dtos(x = as.Date(date(), "\%a \%b \%d \%H:\%M:\%S \%Y"))
}
\arguments{
\item{x}{a varaiable coersable to Date}
}
\description{
Convert Dates into strings with format YYYYMMDD
}
\examples{
dtos()
dtos('2013-01-31')
dtos('01/31/2013')
\dontrun{
dtos('01-31-2013') # non-stanard format
dtos(c('2013-01-31','01/31/2013')) # inconsistent format
}
}

