% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dup}
\alias{dup}
\title{Duplicate a character string n times}
\usage{
dup(x, n)
}
\arguments{
  \item{x}{Input character string}

  \item{n}{A numeric vector (number of times to duplicate)}
}
\value{
A character string
}
\description{
Duplicate a character string n times
}
\examples{
dup("#", 1:10)
}

