% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date-functions.r
\name{easyDateTime}
\alias{easyDateTime}
\title{Coersion to date format.}
\usage{
easyDateTime(x)
}
\arguments{
\item{x}{a varaiable coersable to Date class}
}
\description{
Coersion to date format. Note that this is *much* slower than converting
to date using as.Date(x,pattern) if you already know the date pattern
}
\details{
Note that \code{libarary(lubridate)} is almost certianly a better way to go.
}
\examples{
easyDateTime('2013-01-31')
easyDateTime('01/31/2013')
\dontrun{
easyDateTime('01-31-2013') # non-stanard format
easyDateTime(c('2013-01-31','01/31/2013')) # inconsistent format
}
}

