% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/threshold.r
\name{ecdf2}
\alias{ecdf2}
\title{Empirical thresholds corresponding to given specificities}
\usage{
ecdf2(x, ref, interpMethod = "proportional", na.rm = FALSE)
}
\arguments{
\item{x}{A vector of marker values}

\item{case}{A logical vector where TRUE indicates that the corresonding marker value in X was
is taken from a 'Case', and FALSE otherwise.}

\item{spec}{A vector specificities for the returned thresholds}

\item{...}{(optional) Additoinal arguments to \code{threshold()}}
}
\value{
A list containing the sensetivities, specificiees, and the corresponding thresholds (cutoffs)
}
\description{
Empirical thresholds corresponding to given specificities.
}
\seealso{
\code{\link{'threshold'}}
}

