% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\name{empty}
\alias{empty}
\title{Test if a string is missing or contains only whitespace}
\usage{
empty(x)
}
\arguments{
\item{x}{A character vetor to test for missing or emtpy strings.}
}
\description{
Test if a string is missing or contains only whitespace.
}

