% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\name{first}
\alias{first}
\alias{last}
\title{The position (indicator) of the first or last value that evaluates to \code{TRUE}}
\usage{
first(x, none = NA)

last(x, none = NA)
}
\arguments{
\item{x}{a vector to be searched}

\item{none}{a default return value if none of the values in X evaluate to \code{TRUE}}
}
\description{
The position (indicator) of the first or last value that evaluates to \code{TRUE}.
}

