\name{generateTags}
\alias{generateTags}
\title{generate random tags for minisequencing}
\usage{
  generateTags(n = 20, size = 20, GC_percent = 60,
    max_rep = 4)
}
\arguments{
  \item{n}{number of tags to generate}

  \item{size}{length of tags}

  \item{GC_percent}{average GC content of tags}

  \item{max_rep}{maximum number of identical bases in a
  row}
}
\description{
  generate random tags for minisequencing
}

