% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Short command to create a styled ggplot histogram from a vector}
\usage{
gghist(
  x,
  bins = 100,
  xlab = glue::glue("Variable: {deparse(substitute(x))}"),
  title = ""
)
}
\arguments{
\item{x}{A vector of data}

\item{bins}{Number of histogram bins}

\item{xlab}{x-axis label. Uses the variable name by default.}

\item{title}{Plot title}
}
\value{
ggplot
}
\description{
Short command to create a styled ggplot histogram from a vector
}
