% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gr.r
\name{gr}
\alias{gr}
\title{Regular expressoin searches of the names of a data frame.}
\usage{
gr(pattern, x = names(data), ic = T, ...)
}
\arguments{
\item{pattern}{a reguralar expression pattern}

\item{x}{an object to be searched. If x is a \code{data.frame} then the names of x are searched}
}
\description{
Regular expressoin searches of the names of a data frame
}

