% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\name{\%not in\%}
\alias{\%not in\%}
\title{The opposite of \%in\%}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.  Long vectors are supported}

\item{table}{vector or NULL: the values to be matched against.  Long vectors are not supported.}
}
\description{
lets face it \code{(!x \%in\% y} is just ugly syntax.
}

