\name{has_command}
\alias{has_command}
\title{On a UNIX-like system, test if an external dependency is available}
\usage{
  has_command(cmd, msg = "")
}
\arguments{
  \item{cmd}{The command to test.}

  \item{msg}{Additional message if the test fails.}
}
\description{
  On a UNIX-like system, test if an external dependency is
  available
}
\seealso{
  Other tests: \code{\link{all_empty}},
  \code{\link{all_scalar}}, \code{\link{are_empty}},
  \code{\link{are_scalar}}, \code{\link{is.empty}},
  \code{\link{is.scalar}}
}

