\name{has_tables}
\alias{\%has_tables\%}
\alias{has_tables}
\title{Check if a database has specified tables}
\usage{
  has_tables(con, tbl)

  %has_tables%(con, tbl)
}
\arguments{
  \item{con}{a connection object}

  \item{tables}{a character vector of table names}
}
\description{
  Check if a database has specified tables
}

