% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/intercolate.r
\name{intercolate}
\alias{intercolate}
\title{Intercolte two items}
\usage{
intercolate(x, y, ...)
}
\arguments{
\item{x,y}{two variables of the same class to be intercolated}
}
\description{
This function allows intercolates 2 variables
}
\examples{
intercolate(paste(1:3),LETTERS[1:3])
intercolate(paste(1:4),LETTERS[1:3])
\\notrun{
intercolate(paste(1:4),LETTERS[1:3])
}
}

