% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.r
\name{inv.col}
\alias{inv.col}
\title{Invert a color}
\usage{
inv.col(x, intensity = 255)
}
\arguments{
\item{x}{the colors to invert}

\item{intensity}{(optional) around which to invert the individual R,G, and B values}
}
\description{
Invert a color.
}

