% Generated by roxygen2 (3.1.0.99): do not edit by hand
\name{is.installed}
\alias{is.installed}
\title{Is an R Packages installed?}
\usage{
is.installed(pkg)
}
\arguments{
  \item{pkg}{Package name as character string.}
}
\description{
Is an R Packages installed?
}
\examples{
is.installed("methods")
}

