\name{listFilesRec}
\alias{listFilesRec}
\title{list files recursively}
\usage{
  listFilesRec(path = ".", dir_pattern = NULL,
    exclude_pattern = NULL, file_pattern = NULL)
}
\arguments{
  \item{path}{character vector of path names}

  \item{dir_pattern}{optional regexp}

  \item{exclude_pattern}{optional regexp}

  \item{file_pattern}{optional regexp}
}
\description{
  list files recursively
}

