\name{Log_to}
\alias{Log_to}
\title{Log a time stamp and a message to file everytime a function is run}
\usage{
Log_to(path, message = "", fn)
}
\arguments{
  \item{path}{path to log file}

  \item{message}{logging message}

  \item{fn}{function}
}
\description{
Log a time stamp and a message to file everytime a function
is run
}

