% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-misc.r
\name{logAxis}
\alias{logAxis}
\title{Plots the axis on the log scale}
\usage{
logAxis(side,at,labels,...)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is to be drawn on. See \code{\link[graphics]{axis}} for details}

\item{at}{the points at which tick-marks are to be drawn. See \code{\link[graphics]{axis}} for details}

\item{labels}{labels for the tick marks. See \code{\link[graphics]{axis}} for details}

\item{...}{additional arguments to \code{\link[graphics]{axis}}}
}
\description{
Add an Axis to a Plot onthe log scale. A wrapper for \code{\link[graphics]{axis}}
which supports the ability to log transforming the default axis labels
which comes in handy when calling plot(log(x),log(y),...) is easier than
calling  plot(x,y,log='xy',...)
}
\seealso{
\code{\link[graphics]{axis}}
}

