% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\name{lu}
\alias{lu}
\title{Number of unique elements of a vector}
\usage{
lu(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{If \code{TRUE} missing values are considered a unique value (Default = \code{FALSE})}
}
\description{
Number of unique elements of a vector
}

