% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/match.rev.R
\name{match.rev}
\alias{match.rev}
\title{Reverse a match index}
\usage{
match.rev(x, len)
}
\arguments{
\item{x}{A match index, often the result of calling `x <- match(a,b)`}

\item{len}{the length of the the retusulting index [i.e. `length(b)`]}
}
\description{
Reverse a match index, which can save time when the index takes a long time to create
}

