% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my r programs.r
\name{matchapply}
\alias{matchapply}
\title{I can't really remember what this function does.}
\usage{
matchapply(x, xfactor, fkey, FUN, na, ...)
}
\description{
I can't really remember what this function does
}

