% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meansd.r
\name{meansd}
\alias{meansd}
\title{Pretty printing of the mean and SD of a variable}
\usage{
meansd(x, ...)
}
\arguments{
\item{x}{a variable to be summarized}
}
\description{
This function allows you to express your love of cats.
}

