% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_csv}
\alias{merge_csv}
\title{Short command to create a styled ggplot histogram from a vector}
\usage{
merge_csv(folder_name, output_filename = "results.csv")
}
\arguments{
\item{folder_name}{Which directory (relative to current directory) contains the CSV files?}

\item{output_filename}{How should the output file be named?}
}
\value{
None
}
\description{
Short command to create a styled ggplot histogram from a vector
}
