% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/multiPoint.r
\name{multiPoints}
\alias{multiPoints}
\title{Plotting with compound symbols}
\usage{
multiPoints(x, y, n = if (!missing(col)) rep(ceiling(sqrt(ncol(col))), 2) else
  rep(ceiling(sqrt(ncol(pch))), 2), col = matrix("black", length(x),
  ncol(pch)), bg = matrix(NA, length(x), ncol(col)), pch = matrix(NA,
  length(x), ncol(col)), xheight = strheight("M") * cex * 1.1,
  xwidth = strwidth("M") * cex * 1.1, ...)
}
\arguments{
\item{x,y}{the coordinates of points in the plot}

\item{n}{= if(!missing(col))}

\item{col}{a matrix of colors with 1 row for each plot point. See \code{\link[graphics]{points}} for details.}

\item{bg}{a matrix of background colors with 1 row for each plot point. See \code{\link[graphics]{points}} for details.}

\item{pch}{a matrix of plotting ‘character’, i.e., symbol to use. See \code{\link[graphics]{points}} for details.}

\item{xheight}{= strheight('M')*cex * 1.1,}

\item{xwidth}{= strwidth('M')*cex * 1.1,}

\item{...}{additional arguments to 'plot'}
}
\description{
Plotting with compound symbols made from a matrix of standard symbols
each with their own characteristics
}
\examples{
n <- c(2,2)
x <- 1:5
y <- 1:5
#-- col <- matrix(rainbow(20),5,4)
col <- rbind(rainbow(4),
			 rainbow(4),
			 rainbow(4),
			 rainbow(4),
			 rainbow(4))

col1 <- col
col1[] <- 'hotpink'
col[,3] <- NA
plot(x,y)
multiPoints(x,y,n = 4,col=col)
multiPoints(x,y,n = 1,col=col)
multiPoints(x,y,n = 2,col=col)
multiPoints(x,y,n = 2,col=col1,pch = matrix(1:4,5,4,byrow=TRUE))
multiPoints(x,y,n = 1,col=col1,pch = matrix(letters[1:4],5,4,byrow=TRUE),cex = 2)

plot(x,y)
multiPoints(x,y,n = 1,col=col,cex = 2)
multiPoints(x,y,n = 1,col=col1,pch = matrix(letters[1:4],5,4,byrow=TRUE),cex = 2)
points(x,y,pch = 15)

#multiPoints(x,y,n = 2,col='black',pch = matrix(1:20,5))
multiPoints(x,y,n = 2,col=col)
}

