\name{new_proj}
\alias{new_proj}
\title{Create a new \href{http://projecttemplate.net/getting_started.html}{ProjectTemplate}
project and open RStudio in it.}
\usage{
new_proj(proj, path = getOption("rmisc.proj"), minimal = FALSE)
}
\arguments{
  \item{proj}{The name of the project directory.}

  \item{path}{Base directory of the project.}

  \item{minimal}{Create a minimal project or a full
  project.}
}
\value{
No value is returned; RStudio is opened in the newly
created project.
}
\description{
Create a new
\href{http://projecttemplate.net/getting_started.html}{ProjectTemplate}
project and open RStudio in it.
}

