% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/npercent.r
\name{npercent}
\alias{npercent}
\alias{npercent.numeric}
\title{Pretty printing of a numerator and percentage}
\usage{
npercent(numerator, ...)

\method{npercent}{numeric}(numerator, denominator = sum(numerator, na.rm =
  na.rm), na.rm = FALSE, digits = 1, join = " ", of = FALSE, N = TRUE,
  perc = TRUE, ...)
}
\arguments{
\item{numerator}{The numerator of a fraction}

\item{...}{additional arguments. (ignored)}

\item{denominator}{The denominator of a fraction. (Default = sum(numerator,na.rm = na.rm))}

\item{na.rm}{If TRUE,  missing values are excluded when calculating the numerator and/or denominator.(Default = FALSE)}

\item{digits}{number of digits included in the percentage. (defatlt = 1)}

\item{join}{a separator between the numerator and the folloing open paren. (defatlt = ' ')}

\item{of}{If True, the denominator is included in the out put text. (default=FALSE)}

\item{N}{If True, the numerator is included in the output text. (default= TRUE)}

\item{perc}{If True, the percentage is included in the output. (default= TRUE)}
}
\description{
Pretty printing of a numerator and percentage.
}
\examples{
npercent(6,9)
npercent(6,9,of=TRUE)
npercent(factor(c('a','b','b','c','c','c')),of=TRUE)
npercent(factor(c('a','b','b','c','c','c')),denominator=12,of=TRUE,N=FALSE)
npercent(sample(c(TRUE,FALSE),10,TRUE),of=TRUE)
npercent(strsplit('finished files are the result of years
of scientific study combined with the experience of many years','')[[1]])
}

