% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my r programs.r
\name{nvl}
\alias{nvl}
\title{Replacement of missing values}
\usage{
nvl(x, val)
}
\arguments{
\item{x}{a vector, possibly with missing values}

\item{val}{replacement values for x}
}
\description{
Replacement of missing values
}
\examples{
x <- letters
(is.na(x) <- !! 1:26 \%\% 2)
nvl(x,'Hi')
nvl(x,LETTERS)
\\notrun{ nvl(x,c('Hello','World')) }
}
\keyword{cats}

