\name{:=}
\alias{:=}
\title{parallel assignment}
\arguments{
  \item{lhs}{left hand side of the operator}

  \item{rhs}{right hand side of the operator}
}
\description{
  parallel assignment
}
\note{
  Pinched from
  \url{http://code.google.com/p/miscell/source/browse/rvalues/rvalues.r}

  usage: c(a,b,c) := c(1,2,3)
}

