% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/group-plot.r
\name{plot.group}
\alias{plot.group}
\title{Plot marker values by group}
\usage{
plot.group(group, y, labels = if (inherits(group, "factor")) levels(group)
  else unique(group), log = "", bg.panel, add = FALSE, axes = TRUE,
  xlab = deparse(substitute(group)), ylab = deparse(substitute(y)), ...)
}
\arguments{
\item{group}{Group membership}

\item{y}{Values to be plotted}

\item{labels}{labels to use for each group. Labels must be
in unique(group) and can be use to re-arrange or filter the display of the data}

\item{log}{a character string which contains "x" if the x axis is to be logarithmic, "y" if the y axis is to be logarithmic and "xy" or "yx" if both axes are to be logarithmic. See \code{\link[graphics]{plot.default}} for details.}

\item{bg.panel}{Colors used as backgrounds for the panels}

\item{add}{if TRUE, figures are added to the current plotting region, otherwise a new plot is created}

\item{axes}{axes to be plotted ('x','y','xy', or logical; default = 'xy'). See \code{\link[graphics]{plot.default}} for details.}

\item{xlab,ylab}{labels for the x and y axes. See \code{\link[graphics]{plot.default}} for details.}

\item{...}{additional arguments to \code{\link[graphics]{plot.default}}.}
}
\value{
a list of the x and y locations of the points within the figure
}
\description{
Plot marker values by group with wobble added
}
\examples{
g <- rep(c('small','medium','large'),times= 5)
y <- rep(1:3,times = 5) + runif(15)
pts <- plot.group(group = g,
			 y = y,
			 col = rep(c('blue','green','red'),times= 5),
			 pch = 1:15,
			 labels = c('large','medium','small'),
			 xlab = 'group',
			 ylab = 'y value',
			 main = 'sample group plot',
			 log = 'y',
			 sub = 'hello world',
			 bg.panel = c('gray','white','dodgerblue'))
points(pts$x, pts$y, cex = 2)
}

