% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-misc.r
\name{plotloc}
\alias{plotloc}
\title{Calculate plot positoins from relative coordatinates}
\usage{
plotloc(x = 0.5, y = 0.5)
}
\arguments{
\item{x,y}{x and y coordinats, usually  in the interval [0,1]}
}
\description{
Calculate plot positoins from relative coordatinates. In relative coordinates
The plot region ranges from 0 to 1 on both axes.  The coordinates returned
are in plot coordinates.
}
\examples{
plot(1:5,1:3)
plotloc(1,1) # the top right corner of the plotting region
plotloc(0,0) # the bottom left corner of the plotting region
}

