% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-misc.r
\name{plotlocator}
\alias{plotlocator}
\title{Reads the position of the graphics cursor when the (first) mouse button is pressed.}
\usage{
plotlocator(n = 1, ...)
}
\arguments{
\item{n}{the maximum number of points to locate.  Valid values start
    at 1.}

\item{...}{additional graphics parameters used if \code{type != "n"}
    for plotting the locations.}
}
\description{
Reads the position of the graphics cursor when the (first) mouse button is pressed.
unlike \code{graphics::locator()}, the values returned are in relative coordinates.
In relative coordinates, The plot region ranges from 0 to 1 on both axes.
}
\examples{
\dontrun{
plotlocator()
}
}
\keyword{cats}

