% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log-logistic-inverse-functions.r
\name{predict.5pl.inv}
\alias{predict.5pl.inv}
\title{Inverse of the 5 parameter log-logistic inverse function.}
\usage{
\method{predict}{5pl.inv}(y, b, c, d, e, f)
}
\arguments{
\item{y}{oberseved readings to back transformed into concentrations}

\item{b,c,d,e,f}{parameters of the log-ligistic function}
}
\description{
Inverse of the 5 parameter log-logistic inverse function.  This is useful for
when applying a standard curve to ELISA or other similar data
}
\seealso{
4 parameter log-logistic inverse function: \code{\link{predict.4pl.inv}}
}

