\name{progress_bar}
\alias{progress_bar}
\title{plots a progress bar}
\usage{
  progress_bar(i, n, width = 4, height = 0.4, inv = FALSE)
}
\arguments{
  \item{i}{variable}

  \item{n}{end point of loop}

  \item{width}{width of bar}

  \item{height}{height of bar}

  \item{inv}{if \code{TRUE} the progress bar runs
  backwards}
}
\description{
  put it in a \code{for(i in seq(n))} loop
}

