% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replace_ext}
\alias{replace_ext}
\title{Replace file extensions}
\usage{
replace_ext(file, replacement = "", sep = "\\\\.", level = 0)
}
\arguments{
\item{replacement}{replacement extension}

\item{file}{The file name(s).}

\item{sep}{specifies the seperator character (default ".").}

\item{level}{How many extensions should be stripped.
The default (0) strips all, 1 strips the last one, 2 strips the last two,
and so on.}
}
\description{
Replace file extensions
}

