% Generated by roxygen2 (3.1.0.99): do not edit by hand
\name{require.all}
\alias{require.all}
\title{Require one or more packages}
\usage{
require.all(...)
}
\arguments{
  \item{...}{Package names (as strings).}
}
\description{
A simple wrapper for \code{\link[base]{require}} that
allows loading multiple packages. Issues a warning if a
package cannot be loaded.
}
\seealso{
\code{\link[base]{require}}
}

