\name{Rmisc-package}
\alias{Rmisc-package}
\alias{Rmisc}
\docType{package}
\title{
Rmisc: Ryan Miscellaneous
}
\description{
The Rmisc library contains many functions useful for data analysis and utility operations.
}
\details{
\tabular{ll}{
Package: \tab Rmisc\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-10-05\cr
License: \tab GPL-3\cr
}
}
\author{
Maintainer: Ryan M. Hope <rmh3093@gmail.com>
}

\keyword{ package }
