% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set Operations.r
\name{union}
\alias{intersect}
\alias{union}
\title{Performs the union of 2 or more vectors}
\usage{
union(...)

intersect(...)
}
\arguments{
\item{...}{vectors to be intersected}
}
\description{
Recursively performs the union of 2 or more vectors.
}
\seealso{
\code{\link[base]{union}} and \code{\link[base]{intersect}} on which these functions are based
}

