\name{stripExt}
\alias{stripExt}
\title{Strip file extensions}
\usage{
  stripExt(file, sep = "\\.", level = 0)
}
\arguments{
  \item{file}{file name(s)}

  \item{sep}{specifies the seperator character (default
  ".").}

  \item{level}{numeric denoting how many extensions should
  be stripped. The default (0) strips all, 1 strips the
  last one, 2 strips the last two, and so on}
}
\description{
  Strip file extensions
}

