% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarize.r
\name{summarize}
\alias{summarize}
\title{Summarize a variable in a nice table}
\usage{
summarize(x, varName = deparse(substitute(x)), lof = F, mof = F,
  lyesonly = F, digits = 1, ...)
}
\arguments{
\item{x}{a variabile to summarized}

\item{varName}{= 'x' # name of the variable x}

\item{mof}{if \code{TRUE}  the number of missing values is included in text summary}

\item{lyesonly}{If \code{TRUE}, the number of false reponses is dropped from the output table (default = F)}

\item{digits}{number of decimals used if the summary statistic is the mean and standard deviation (default = c(1,1))}

\item{love}{Do you love cats? Defaults to TRUE.}

\item{lof=F}{if \code{TRUE}  the denomiator is included in the percntage
[example:\code{"15 (15\% of of 100)"}]}

\item{digits.meansd}{deciaml precision when the summary statistic is
the mean and sd of x}
}
\description{
Summarize a variable in a nice table that is compatibile with summaries produced by \code{\link{twoByN}}
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\seealso{
\code{\link{twoByN}} for summarizeing a variable between two groups
}
\keyword{cats}

