% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\name{summary.anova}
\alias{summary.anova}
\title{Summary Statisics for an Anova Object}
\usage{
\method{summary}{anova}(object, ...)
}
\arguments{
\item{object}{an object of class 'anova'}

\item{...}{additional arguments. (ignored)}
}
\description{
Summary Statisics for an Anova Object.
}
\examples{
summary(aov(y ~ x ,data=data.frame(y = rnorm(50),x=gl(5,10,50,LETTERS[1:5]))))
}

