% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my-r-programs.r
\docType{data}
\name{temp.pdf}
\alias{temp.csv}
\alias{temp.pdf}
\title{Pointers to a temporary files}
\format{\preformatted{ chr "\\\\\\\\home.fhcrc.org\\\\jthorpe/temp\\\\temp.pdf"
}}
\usage{
temp.pdf

temp.csv
}
\description{
Pointers to a temporary files.
}
\keyword{datasets}

