% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tochar.r
\name{tochar}
\alias{tochar}
\title{Convert object to pretty strings}
\usage{
tochar(x, ...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Convert objects to strings with more contol than \code{\link[base]{as.character}()}
and less confusion than \code{\link[base]{format}()}
}
\examples{
tochar(0.00123,digits=4)
as.character(round(0.00123,digits=4)) #same as above

tochar(123,digits=4)
as.character(round(123,digits=4))# no significant digits

tochar(0.000123,digits=4)
as.character(round(0.000123,digits=4)) #not pretty

tochar(0.0000123,digits=4)
as.character(round(0.0000123,digits=4)) #no significant digits

\\notrun{
tochar(sin)# cant coerce a function to string...
}
}

