\name{trim}
\alias{trim}
\title{Trim elements}
\usage{
  trim(x, trim = "\\s+", fixed = FALSE)
}
\arguments{
  \item{x}{A vector.}

  \item{trim}{A pattern to trim from vector elemts}

  \item{fixed}{treat \code{trim} as a fixed string}
}
\value{
  A vector with trimmed strings
}
\description{
  Trim elements
}

