% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twoByN.r
\name{twobyn}
\alias{twobyn}
\title{Summarize balance of a variable between two groups}
\usage{
twobyn(x, ltwo, variable.desc = deparse(substitute(x)), ...)
}
\arguments{
\item{x}{a variabile to be compared between levels of `ltwo`}

\item{ltwo}{# a logical variable for the second dimension of the resulting table}

\item{variable.desc}{= deparse(substitute(x)) # name of the variable x}

\item{...}{# additional arguments to npercent methods}

\item{digits.p.value}{precision of p-value (default=4)}

\item{poisson}{= FALSE # should INTEGER values be treated as counts. If True,
p-values are calculated using poisson regresson?
SPECIFIC TO NUMERIC, INTEGER VARIABLES}

\item{digits.meansd=1}{deciaml precision when the summary statistic is
the mean and sd of x

specific to logical variables}

\item{lyesonly}{if TRUE the number of false reponses is dropped from the output table}

\item{na.as.category}{If TRUE, missing values are treated as if they are a
separate caterory for variables that are compared as categorical variables
NOT APPLICABLE FOR WHEN X IS NUMERIC OR  INTEGER.}

\item{na.inline}{If \code{TRUE}  missing values are reported in the bottom row,
\strong{otherwise} in an additional column is added to the table with summaries
 of the missing values (na.inline=FALSE).  (default = TRUE)}

\item{na.always}{If \code{TRUE}, a count of the missing values always included in the summary table (default = FALSE)?}

\item{na.never}{If \code{TRUE}, missing values are ignored (default = FALSE)}
}
\description{
Summarize balance of a variable between two groups
}
\seealso{
\code{\link{summarize}} for summarizeing a variable
}

