% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/twobytwo.r
\name{twoByTwo}
\alias{twoByTwo}
\title{Calculate the contents of a 2 by 2 table from any 4 (independent) values.}
\usage{
twoByTwo(...)
}
\arguments{
\item{a,b,ab,cd,acd,abcd,etc.}{the sum of 1 or more cells of the matrix \tabular{cc}{a \tab c \cr b \tab d }
with the parameter names being alphabetical lists of the cells which they sum.}

\item{odds.ratio}{the odds ratio (\code{(a/b)/(c/d)})}
}
\description{
Calculate the contents, margins, and/or Odds Ratio from any 4 (independent) values of a two by two table.
}
\details{
Every onece in a while, a the contents of a two-by-two table need to
be calculated based on the margins, odds ratios, and/or sum of the entries, which
can be a nuisance to calculate by hand.  Hence \code{twoByTwo} calculates a two-by-two matrix, from any
combination of 4 parameters which fully specify the contents of a two by two table.
parameters are named based on the following table :
\tabular{cc}{a \tab c \cr b \tab d }
}
\examples{
# most trivial of all, same as matrix(1:4,2)
twoByTwo(a= 1,b = 2,c = 3,d = 4)

# still trivial
twoByTwo(a= 1,ab=2,abc=3,abcd=4)
twoByTwo(c= 1,ab=2,ac=3,abcd=4)

# not too difficult, but anoying to have to calculate by hand
twoByTwo(odds.ratio = 1,ab=2,ac=3,abcd=5)
twoByTwo(odds.ratio = 1.001,ab=2,ac=3,abcd=5)

\\notrun{
	# error 'n' is not a valid parameter, use 'abcd' instead
	twoByTwo(a= 1,ab=2,ac=3,n=4)

	# error: the parameter ab appears twice
	twoByTwo(ab= 1,ab=2,ac=3,abcd=4)

	# error: insufficient parameters
	twoByTwo(odds.ratio = 1.001,c=2,ac=3,abcd=5)
}
}

