% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/typed.r
\name{typed}
\alias{typed}
\title{Typed: Pretty printing to the console}
\usage{
typed(x)
}
\arguments{
\item{x}{a variable to print to the console}
}
\description{
Pretty printing of variables, which is particulary useful if the printed text will be used
in a script or processed in a text editor.
}
\examples{
typed(1:4)
typed(sample(c(TRUE,FALSE),5,TRUE))
typed(c('one','two','three'))
}

