% Generated by roxygen2 (4.0.0): do not edit by hand
\name{update_github}
\alias{update_github}
\title{A quick installer for my R packages on GitHub}
\usage{
update_github()
}
\description{
Prompts for which of the following packages you want to install:
\code{rmisc}, \code{Rentrez}, \code{biofiles}, \code{blastr},
\code{ncbi}.
}

