\name{update_packages}
\alias{update_packages}
\title{Package updater}
\usage{
  update_packages(destdir = "~/R/Packages")
}
\arguments{
  \item{destdir}{Where the source code goes}
}
\description{
  Check for outdated packages, save the source to
  \code{destdir}, extract the packages in \code{destdir},
  and install
}

