% Generated by roxygen2 (4.0.1): do not edit by hand
\name{usplit}
\alias{usplit}
\title{unlist(strsplit(x, split, ...))}
\usage{
usplit(x, split, ...)
}
\arguments{
\item{x}{Character vector to be split.}

\item{split}{The regexp to split on.}

\item{\dots}{Arguments passed on to \code{\link{strsplit}}.}
}
\description{
unlist(strsplit(x, split, ...))
}
\examples{
usplit("a.b.c", ".", fixed = TRUE)
}

