% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/venn.r
\name{venn}
\alias{venn}
\title{Plot Venn Diagrams}
\usage{
venn(data, main = "", labels = dimnames(data)[[2]], ntext = "All Others",
  add = F, va = 0, col, col.txt, lty, prefix = "", suffix = "")
}
\arguments{
\item{data}{a \code{data.frame} containing 1 to 4 indicator (logical) variables}

\item{main}{the title of the figure}

\item{labels}{labels for the indicator variables (defaults to names(data))}

\item{ntext}{a label to describe objects that are \code{FALSE} for every indicator variable (default = 'All Others')}

\item{add}{If \code{TRUE}, the digram is added to an existing plot}

\item{va}{a typesetting parameter}

\item{col}{symbol color}

\item{col.txt}{text color}

\item{lty}{line type}

\item{prefix}{I can't remember}

\item{suffix}{I can't remember}
}
\description{
Make Venn Diagrams to Summarize 1 to 4 indicator (logical) variables
}
\examples{
venn( data.frame(A = rnorm(1000)>0
					 ,B = rnorm(1000)>0
					 ,C = rnorm(1000)>0
					 ,D = rnorm(1000)>0
					 ))
}

