\name{wrapAln}
\alias{wrapAln}
\title{Wrap alignment}
\usage{
  wrapAln(seq1, ..., prefix = c(""), suffix = c(""),
    start = c(1), reverse = c(FALSE), sep = 2)
}
\arguments{
  \item{seq1}{First sequence.}

  \item{...}{More sequences}

  \item{prefix}{Character vector of text going before the
  alignment}

  \item{suffix}{Text going after the alignment}

  \item{start}{Numeric vector of starting positions of the
  sequences.}

  \item{reverse}{Any reversed sequences}

  \item{sep}{Space between prefix/suffix and position
  numbering.}
}
\description{
  Wrap alignment
}

