\name{blanks}
\alias{blanks}
\title{Create blank strings with a given number of characters}
\usage{
  blanks(n)
}
\arguments{
  \item{n}{A numeric vector (number of times to duplicate)}
}
\value{
  A character vector
}
\description{
  Create blank strings with a given number of characters
}
\examples{
blanks(10)
}

