% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/colors.r
\name{colorchoose}
\alias{colorchoose}
\title{Choose from among the named colors in R
Choose from among the named colors in R using a graphical interface}
\usage{
colorchoose(n = 1, alpha, term = F)
}
\arguments{
\item{alpha}{(optional) an integer between 1 and 255 with the alpha chanel}

\item{N}{the number of colors to chose}
}
\description{
Choose from among the named colors in R
Choose from among the named colors in R using a graphical interface
}
\examples{
cat_function()
}

