\name{Curry}
\alias{Curry}
\title{Curry a function}
\usage{
  Curry(FUN, ...)
}
\arguments{
  \item{FUN}{Function to curry.}

  \item{...}{Arguments}
}
\description{
  Pinched from
  \href{https://github.com/hadley/devtools/wiki/First-class-functions}{here}.
}
\author{
  Hadley Wickham <h.wickham@gmail.com>
}

