% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date-functions.r
\name{easyDateTime}
\alias{easyDateTime}
\title{Coersion to date format.}
\usage{
easyDateTime(x)
}
\arguments{
\item{x}{a varaiable coersable to Date}
}
\description{
Coersion to date format. Note that this is *much* slower than converting
to date using as.Date(x,pattern) if you already know the date pattern
}
\examples{
as.date('2013-01-31')
as.date('01-31-2013')
as.date('01/31/2013')
as.date(c('2013-01-31','01/31/2013'))
}

