% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/threshold.r
\name{ecdf2}
\alias{ecdf2}
\title{Empirical thresholds corresponding to given specificities}
\usage{
ecdf2(x, ref, spec, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of marker values}

\item{ref}{an optional reference group.}

\item{spec}{A vector specificities for the returned thresholds}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A list containing the sensetivities, specificiees, and the corresponding thresholds (cutoffs)
}
\description{
Empirical thresholds corresponding to given specificities.
}
\seealso{
\code{\link{threshold}}
}

