% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-misc.r
\name{logAxis}
\alias{logAxis}
\title{Plots the axis on the log scale}
\usage{
logAxis(side, at = log(labels), labels = if ((fold = diff(par("usr")[if
  (side\%\%2) 1:2 else 3:4]/log(10))) < 1.4) {     i <- 0     usr <-
  par("usr")[if (side\%\%2)          1:2     else 3:4]     while (TRUE) {       
   if (i == 20)              stop("range too small")         i <- i + 1        
  if (ceiling(exp(usr)[1]) < floor(exp(usr)[2]))              break         else
  usr <- usr * 10     }     seq.int(ceiling(exp(usr)[1]) * (10^i),
  floor(exp(usr)[2]) * (10^i), length.out = 5)/10^i } else axTicks(side, log =
  TRUE, usr = (par("usr")[if (side\%\%2) 1:2 else 3:4]/log(10)), axp =
  c(10^ceiling(par("usr")[if (side\%\%2) 1 else 3]/log(10)),
  10^floor(par("usr")[if (side\%\%2) 2 else 4]/log(10)), n = if ((fold =
  diff(par("usr")[if (side\%\%2) 1:2 else 3:4]/log(10))) < 1.4) -1 else if
  ((fold = diff(par("usr")[if (side\%\%2) 1:2 else 3:4]/log(10))) > ((width =
  par("pin")[if (side\%\%2) 1 else 2]))) 1 else if ((fold = diff(par("usr")[if
  (side\%\%2) 1:2 else 3:4]/log(10))) * 1.3 > ((width = par("pin")[if
  (side\%\%2) 1 else 2]))) 2 else 3),      nintLog = c(if ((fold =
  diff(par("usr")[if (side\%\%2) 1:2 else 3:4]/log(10))) < 1.4) 5 else if ((fold
  = diff(par("usr")[if (side\%\%2) 1:2 else 3:4]/log(10))) > ((width =
  par("pin")[if (side\%\%2) 1 else 2]))) floor(fold + 1) else if ((fold =
  diff(par("usr")[if (side\%\%2) 1:2 else 3:4]/log(10))) * 1.3 > ((width =
  par("pin")[if (side\%\%2) 1 else 2]))) floor(fold * 2) else floor(fold * 3))),
  ...)
}
\arguments{
\item{side}{an integer specifying which side of the plot the axis is
    to be drawn on.  The axis is placed as follows: 1=below,
    2=left, 3=above and 4=right.}

\item{at}{the points at which tick-marks are to be drawn.  Non-finite
    (infinite, \code{NaN} or \code{NA}) values are omitted.  By default
    (when \code{NULL}) tickmark locations are computed, see
    \sQuote{Details} below.}

\item{labels}{this can either be a logical value specifying whether
    (numerical) annotations are to be made at the tickmarks, or a
    character or expression vector of labels to be placed at the
    tickpoints.  (Other objects are coerced by \code{\link{as.graphicsAnnot}}.)
    If this is not logical, \code{at} should also be supplied and of the
    same length.  If \code{labels} is of length zero after coercion,
    it has the same effect as supplying \code{TRUE}.}

\item{...}{other \link{graphical parameters} may also be passed as
    arguments to this function, particularly, \code{cex.axis},
    \code{col.axis} and \code{font.axis} for axis annotation, \code{mgp}
    and \code{xaxp} or \code{yaxp} for positioning, \code{tck} or
    \code{tcl} for tick mark length and direction, \code{las} for
    vertical/horizontal label orientation, or \code{fg} instead of
    \code{col}, and \code{xpd} for clipping.  See \code{\link{par}} on these.

    Parameters \code{xaxt} (sides 1 and 3) and \code{yaxt} (sides 2 and
    4) control if the axis is plotted at all.

    Note that \code{lab} will partial match to argument
    \code{labels} unless the latter is also supplied.  (Since the
    default axes have already been set up by \code{\link{plot.window}},
    \code{lab} will not be acted on by \code{axis}.)}
}
\description{
Add an Axis to a Plot onthe log scale. A wrapper for \code{\link[graphics]{axis}}
which supports the ability to log transforming the default axis labels
which comes in handy when calling plot(log(x),log(y),...) is easier than
calling  plot(x,y,log='xy',...)
}
\seealso{
\code{\link{axis}}
}

