% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Maybe}
\alias{Maybe}
\title{Maybe call a function}
\usage{
Maybe(fn)
}
\arguments{
\item{fn}{A function.}
}
\value{
A function.
}
\description{
If the argument(s) to a function are missing or \code{NULL},
return \code{NULL}; otherwise apply the function.
}

