% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/group plot.r
\name{plot.group}
\alias{plot.group}
\title{Plot marker values by group}
\usage{
\method{plot}{group}(group, y, labels = if (inherits(group, "factor"))
  levels(group) else unique(group), log = "", bg.panel, add = F, axes = T,
  xlab = deparse(substitute(group)), ylab = deparse(substitute(y)), ...)
}
\arguments{
\item{group}{Group membership}

\item{y}{Values to be plotted}

\item{labels}{labels to use for each group. Labels must be
in unique(group) and can be use to re-arrange or filter the display of the data}

\item{bg.panel}{Colors used as backgrounds for the panels}

\item{axes}{Axes to be plotted ('x','y','xy', or logical; default = 'xy')}

\item{...}{additional arguments to plot()}
}
\value{
a list of the x and y locations of the points within the figure
}
\description{
Plot marker values by group with wobble added
}
\examples{
g <- ep(c('small','medium','large'),times= 5)
y <- rep(1:3,times = 5) + runif(15),
pts <- plot.group(group = g,
			 y = y,
			 col = rep(c('blue','green','red'),times= 5),
			 pch = 1:15,
			 labels = c('large','medium','small'),
			 xlab = 'group',
			 ylab = 'y value',
			 main = 'sample group plot',
			 log = 'y',
			 sub = 'hello world',
			 bg.panel = c('gray','white','dodgerblue') #' 			)
points(pts$x,
		pts$y,
		cex = 2)
}

