% Generated by roxygen2 (3.1.0.99): do not edit by hand
\name{Sequence}
\alias{Sequence}
\title{Compose functions}
\usage{
Sequence(...)
}
\arguments{
  \item{\dots}{The functions to be composed.}
}
\description{
Same as \code{\link{Compose}}, except that it applies the
functions in argument-list order.
}

