% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my r programs.r
\name{trim}
\alias{trim}
\title{Trim leading and trailing white space from character vectors}
\usage{
trim(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
x, trimmed of leading and trailing white space
}
\description{
Trim leading and trailing white space from character vectors.
}
\examples{
trim(c('  hello world\\t'))
}

