% Generated by roxygen2 (3.1.0.99): do not edit by hand
\name{update_github}
\alias{update_github}
\title{A quick installer for my R packages on GitHub}
\usage{
update_github()
}
\description{
Prompts for which of the following packages you want to
install: \code{rmisc}, \code{Rentrez}, \code{biofiles},
\code{blastr}, \code{ncbi}.
}

