% Generated by roxygen2 (3.1.0.99): do not edit by hand
\name{wrap}
\alias{wrap}
\title{Wrap elements}
\usage{
wrap(x, wrap = "\\"")
}
\arguments{
  \item{x}{A vector.}

  \item{wrap}{A string to wrap around vector elements}
}
\value{
a vector with wrapped strings
}
\description{
Wrap elements
}
\examples{
wrap(c("abc", "xyz"), "---")
}

